move CurveTo to separate file
