+++ core/src/commonMain/kotlin/dev/lennartegb/vec2compose/core/commands/CurveTo.kt
@@ -1,0 +1,14 @@
+package dev.lennartegb.vec2compose.core.commands
+
+/**
+ * Curve to command indicated by c/C
+ */
+data class CurveTo(
+    val x1: Float,
+    val y1: Float,
+    val x2: Float,
+    val y2: Float,
+    val x3: Float,
+    val y3: Float,
+    override val isAbsolute: Boolean
+) : Command
