+++ core/src/commonTest/kotlin/dev/lennartegb/vec2compose/core/PathParserTest.kt
@@ -1,7 +1,7 @@
 package dev.lennartegb.vec2compose.core
 
 import dev.lennartegb.vec2compose.core.commands.Close
-import dev.lennartegb.vec2compose.core.commands.Command
+import dev.lennartegb.vec2compose.core.commands.CurveTo
 import dev.lennartegb.vec2compose.core.commands.LineTo
 import dev.lennartegb.vec2compose.core.commands.MoveTo
 import kotlin.test.Test
@@ -18,7 +18,7 @@ internal class PathParserTest {
             expected = listOf(
                 MoveTo(x = 27.05f, y = 24.55f, isAbsolute = true),
                 LineTo(x = 12.15f, y = 19.65f, isAbsolute = true),
-                Command.CurveTo(11.85f, 19.35f, 11.483f, 19.2f, 11.05f, 19.2f, isAbsolute = true)
+                CurveTo(11.85f, 19.35f, 11.483f, 19.2f, 11.05f, 19.2f, isAbsolute = true)
             )
         )
     }
@@ -32,7 +32,7 @@ internal class PathParserTest {
             expected = listOf(
                 MoveTo(x = 27.05f, y = 24.55f, isAbsolute = true),
                 LineTo(x = 12.15f, y = 19.65f, isAbsolute = true),
-                Command.CurveTo(11.85f, 19.35f, 11.483f, 19.2f, 11.05f, 19.2f, isAbsolute = true),
+                CurveTo(11.85f, 19.35f, 11.483f, 19.2f, 11.05f, 19.2f, isAbsolute = true),
                 Close
             )
         )
